+++ ChangeLog.md
@@ -1,4 +1,24 @@
 # Changelog
+## Version 0.3.5
+**From version v0.3.4 to v0.3.5 some output variables are renamed and the name is taken over by new added variables with different meaning. When upgrading, read the Changelog carefully**
+
+* masking rain by *flag_surface_rain* and *flag_rain_ze* is now applied only if precipitation is detected in the lowest range-gate
+* **vsplot:**
+  * More customizations options, e.g., fontsize and data selection
+  * rename *vsplot.flag_surface_rain* to *vsplot.flag_rain*, as it plots now the combination of *flag_surface_rain* and *flag_rain_ze*
+  * rename *vsplot.quicklook_flag_virga* to *vsplot.quicklook_virga_mask*
+  * rename *vsplot.plot_flag_virga* to *vsplot.plot_virga_mask*
+  * rename *vsplot.plot_flag_cloud* to *vsplot.plot_cloud_mask*
+  * rename *vsplot.plot_flag_ze* to *vsplot.plot_ze_mask*
+* **Output dataset:**
+  * rename *flag_virga* to *mask_virga*
+  * rename *flag_cloud* to *mask_cloud*
+  * rename *flag_virga_layer* to *mask_virg_layera*
+  * rename *flag_cloud_layer* to *mask_cloud_layer*
+  * added *flag_lowest_rg_rain*, with the flag if rain is detected at lowest range-gate from the radar reflectivity
+  * added *flag_virga* and *flag_virga_layer* as a flag if virga is detected at a time-step and layer
+  * added *flag_cloud* and *flag_cloud_layer* as a flag if virga is detected at a time-step and layer
+  * added *number_cloud_layers*
 
 ## Version 0.3.4